+++ src/components/inputs/code-input/code-input.model.ts
@@ -1,0 +1,5 @@
+export enum CodeInputCase {
+  UPPERCASE = 'UPPERCASE',
+  LOWERCASE = 'LOWERCASE',
+  DEFAULT = 'DEFAULT',
+}