+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release notes
 
-## Unreleased
+## 0.5.0
 
 ### Features